# swagger.yaml
basePath: /api/v1
consumes:
- application/json
...
host: localhost
info:
  description: |-
    The purpose of this application is to provide the client with functionality to define users, items for auctions,
    and ways to place bids on those items.
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
  termsOfService: There are no terms of service at the moment. Use at your own risk,
    we take no responsibility.
  title: AuctionHouse API.
  version: 0.1.0
paths:
  /api/v1/auctions/bids:
    get:
      description: This will retrieve the highest bid for all items.
      operationId: getHighestBidsRequest
      parameters:
      - description: Expected to be "Bearer <auth_token>"
        in: header
        name: Authentication
        type: string
      produces:
      - application/json
      responses:
        "200":
          $ref: '#/responses/getHighestBidResponse'
      schemes:
      - http
      security:
      - api_key: []
      summary: Retrieves the highest bid for all items.
      tags:
      - Auctions
  /api/v1/auctions/bids/{itemName}:
    get:
      description: This will retrieve the highest bid for the specified item.
      operationId: getHighestBidRequest
      parameters:
      - description: Expected to be "Bearer <auth_token>"
        in: header
        name: Authentication
        type: string
      - in: path
        name: itemName
        required: true
        type: string
        x-go-name: ItemName
      produces:
      - application/json
      responses:
        "200":
          $ref: '#/responses/getHighestBidResponse'
        "404":
          $ref: '#/responses/errorMessage'
      schemes:
      - http
      security:
      - api_key: []
      summary: Retrieves the highest bid for the specified item.
      tags:
      - Auctions
    post:
      consumes:
      - application/json
      description: |-
        This will place a bid on the specified item. The user is identified by the authentication token. This is only
        available for Bidder users.
      operationId: postBidRequest
      parameters:
      - description: Expected to be "Bearer <auth_token>"
        in: header
        name: Authentication
        type: string
      - in: path
        name: itemName
        required: true
        type: string
        x-go-name: ItemName
      - in: body
        name: Body
        schema:
          properties:
            bidAmount:
              description: The amount to bid on the item for.
              format: int64
              type: integer
              x-go-name: BidAmount
          required:
          - bidAmount
          type: object
      produces:
      - application/json
      responses:
        "201":
          $ref: '#/responses/noBody'
        "400":
          $ref: '#/responses/errorMessage'
        "404":
          $ref: '#/responses/errorMessage'
      schemes:
      - http
      security:
      - api_key: []
      summary: Makes a new bid on an item.
      tags:
      - Auctions
  /api/v1/auctions/items:
    get:
      description: This will retrieve all items from storage.
      operationId: getItemsRequest
      parameters:
      - description: Expected to be "Bearer <auth_token>"
        in: header
        name: Authentication
        type: string
      produces:
      - application/json
      responses:
        "200":
          $ref: '#/responses/getItemsResponse'
      schemes:
      - http
      security:
      - api_key: []
      summary: Gets all items that are currently stored in the system.
      tags:
      - Auctions
    post:
      consumes:
      - application/json
      description: This will create a new item available for being auctioned. This
        route is only available to Admin users.
      operationId: postItemRequest
      parameters:
      - description: Expected to be "Bearer <auth_token>"
        in: header
        name: Authentication
        type: string
      - in: body
        name: Body
        schema:
          properties:
            description:
              description: Description of the item.
              type: string
              x-go-name: Description
            image:
              description: Reference to the image source.
              type: string
              x-go-name: ImageRef
            name:
              description: Name used to identify the item later.
              type: string
              x-go-name: Name
          required:
          - name
          type: object
      produces:
      - application/json
      responses:
        "201":
          $ref: '#/responses/noBody'
        "400":
          $ref: '#/responses/errorMessage'
      schemes:
      - http
      security:
      - api_key: []
      summary: Creates a new item for auction.
      tags:
      - Auctions
  /api/v1/auctions/items/{itemName}:
    delete:
      description: This will delete an existing item. This route is only available
        to Admin users.
      operationId: deleteItemRequest
      parameters:
      - description: Expected to be "Bearer <auth_token>"
        in: header
        name: Authentication
        type: string
      - in: path
        name: itemName
        required: true
        type: string
        x-go-name: ItemName
      produces:
      - application/json
      responses:
        "200":
          $ref: '#/responses/noBody'
        "404":
          $ref: '#/responses/errorMessage'
      schemes:
      - http
      security:
      - api_key: []
      summary: Deletes an item from the server.
      tags:
      - Auctions
    get:
      description: This will retrieve a item from storage based on the name.
      operationId: getItemRequest
      parameters:
      - description: Name of the item.
        in: path
        name: itemName
        required: true
        type: string
        x-go-name: Name
      - description: Expected to be "Bearer <auth_token>"
        in: header
        name: Authentication
        type: string
      produces:
      - application/json
      responses:
        "200":
          $ref: '#/responses/getItemResponse'
        "404":
          $ref: '#/responses/noBody'
      schemes:
      - http
      security:
      - api_key: []
      summary: Gets the item specified by the name.
      tags:
      - Auctions
    put:
      consumes:
      - application/json
      description: This will update an existing item available for being auctioned.
        This route is only available to Admin users.
      operationId: putItemRequest
      parameters:
      - description: Expected to be "Bearer <auth_token>"
        in: header
        name: Authentication
        type: string
      - in: path
        name: itemName
        required: true
        type: string
        x-go-name: ItemName
      - in: body
        name: Body
        schema:
          properties:
            description:
              description: Description of the item.
              type: string
              x-go-name: Description
            image:
              description: Reference to the image source.
              type: string
              x-go-name: ImageRef
            name:
              description: Name used to identify the item later.
              type: string
              x-go-name: Name
          required:
          - name
          type: object
      produces:
      - application/json
      responses:
        "200":
          $ref: '#/responses/noBody'
        "400":
          $ref: '#/responses/errorMessage'
        "404":
          $ref: '#/responses/errorMessage'
      schemes:
      - http
      security:
      - api_key: []
      summary: Updates fields for an item.
      tags:
      - Auctions
...
produces:
- application/json
responses:
  errorMessage:
    description: Response when there is an error with the request.
    schema:
      properties:
        message:
          description: The human readable message to the client.
          type: string
          x-go-name: Message
      type: object
  getHighestBidResponse:
    description: Contains data about the bid, what the item is, and who made it.
    schema:
      properties:
        bidAmount:
          description: The amount of money being bid for this item.
          format: int64
          type: integer
          x-go-name: BidAmount
        bidder:
          description: The user who bid on the item.
          properties:
            displayName:
              description: The display name of the user.
              type: string
              x-go-name: DisplayName
            username:
              description: The username of the user.
              type: string
              x-go-name: Username
          required:
          - username
          type: object
          x-go-name: Bidder
        item:
          description: The item being bid on.
          properties:
            description:
              description: The description of the item.
              type: string
              x-go-name: Description
            image:
              description: The reference to the image source.
              type: string
              x-go-name: ImageRef
            name:
              description: The name of the item.
              type: string
              x-go-name: Name
          required:
          - name
          type: object
          x-go-name: Item
      required:
      - bidAmount
      - item
      - bidder
      type: object
...
  getItemResponse:
    description: Contains data about the item and how to identify them.
    schema:
      properties:
        description:
          description: The description of the item.
          type: string
          x-go-name: Description
        image:
          description: The reference to the image source.
          type: string
          x-go-name: ImageRef
        name:
          description: The name of the item.
          type: string
          x-go-name: Name
      required:
      - name
      type: object
  getItemsResponse:
    description: Contains data about the item and how to identify them.
    schema:
      items:
        properties:
          description:
            description: The description of the item.
            type: string
            x-go-name: Description
          image:
            description: The reference to the image source.
            type: string
            x-go-name: ImageRef
          name:
            description: The name of the item.
            type: string
            x-go-name: Name
        required:
        - name
        type: object
      type: array
...
  noBody:
    description: Response contains no body.
...
schemes:
- http
securityDefinitions:
  api_key:
    in: header
    name: KEY
    type: apiKey
swagger: "2.0"